tests: share one FakeModel stub in conftest
Four test files carried byte-identical FakeModel definitions; the
compaction-invariants suite added the fourth copy. Hoist the stub into
tests/conftest.py next to the other shared fakes and import it
everywhere.
